Move recon_resize() to _utils.py
